feat : edit password , only allow to update account after 48 hours
- edit password api
- only allow to update account and change password after 48 hours from the last time

+++ src/main/java/com/example/banking_system/domain/account/service/domain/AccountService.java
@@ -1,6 +1,7 @@
 package com.example.banking_system.domain.account.service.domain;
 
 import com.example.banking_system.common.exception.ValidationException;
+import com.example.banking_system.domain.account.dto.EditPasswordRequest;
 import com.example.banking_system.domain.account.dto.GetAccountResponse;
 import com.example.banking_system.domain.account.entity.Account;
 import com.example.banking_system.domain.account.entity.BusinessAccount;
@@ -10,10 +11,14 @@ import com.example.banking_system.domain.account.mapper.AccountMapper;
 import com.example.banking_system.domain.account.service.query.AccountQueryService;
 import com.example.banking_system.common.utility.JwtUtil;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.security.SecureRandom;
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
 
 @Service
 @RequiredArgsConstructor
@@ -22,6 +27,10 @@ public class AccountService {
     private final AccountMapper accountMapper;
     private final SecureRandom secureRandom;
     private final JwtUtil jwtUtil;
+    private final PasswordEncoder passwordEncoder;
+
+    @Value("${value.environment}")
+    private String ENVIRONMENT;
 
     @Transactional(readOnly = true)
     public GetAccountResponse get() {
@@ -51,9 +60,35 @@ public class AccountService {
 
     @Transactional
     public void delete() {
+        if(ENVIRONMENT.equals("prod")) {
+            throw new ValidationException("Account deletion is not allowed in production environment");
+        }
+
         final String username = jwtUtil.getUsername();
         Account account = accountQueryService.findByUsername(username);
         accountQueryService.delete(account);
+    }
+
+    @Transactional
+    public void editPassword(EditPasswordRequest request) {
+        final String username = jwtUtil.getUsername();
+        Account account = accountQueryService.findByUsername(username);
+
+        boolean changedPasswordIn48Hours = account.getUpdatedPasswordAt() != null
+                && account.getUpdatedPasswordAt().plus(48, ChronoUnit.HOURS).isAfter(Instant.now());
+
+        if (changedPasswordIn48Hours) {
+            throw new ValidationException("password can only be changed once every 48 hours");
+        }
+
+        if (!passwordEncoder.matches(request.getCurrentPassword(), account.getPassword())) {
+            throw new ValidationException("password is incorrect");
+        }
+
+        Instant now = Instant.now();
+        account.setPassword(passwordEncoder.encode(request.getNewPassword()));
+        account.setUpdatedPasswordAt(now);
+        accountQueryService.save(account);
     }
 
     public String generateAccountNumber() {
